Update test_models to use new parser.

+++ onmt/encoders/audio_encoder.py
@@ -99,7 +99,7 @@ class AudioEncoder(EncoderBase):
         src = src.transpose(0, 1).transpose(0, 3).contiguous() \
                  .view(t, batch_size, nfft)
         orig_lengths = lengths
-        lengths = lengths.view(-1).tolist()
+        lengths = lengths.float().view(-1).tolist()
 
         for l in range(self.enc_layers):
             rnn = getattr(self, 'rnn_%d' % l)
@@ -112,7 +112,7 @@ class AudioEncoder(EncoderBase):
             t, _, _ = memory_bank.size()
             memory_bank = memory_bank.transpose(0, 2)
             memory_bank = pool(memory_bank)
-            lengths = [int(math.floor((length - stride)/stride + 1))
+            lengths = [int(math.floor((length - stride) / stride + 1))
                        for length in lengths]
             memory_bank = memory_bank.transpose(0, 2)
             src = memory_bank
